pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "bdi-llm-formal-ver"
version = "0.1.0"
description = "PNSV: Pluggable Neuro-Symbolic Verification for LLM-generated plans"
requires-python = ">=3.10"

dependencies = [
    "dspy-ai>=2.4",
    "networkx>=3.0",
    "pydantic>=2.0",
    "litellm>=1.0",
    "openai>=1.0",
    "python-dotenv>=1.0",
    "z3-solver>=4.12",
    "mcp>=0.1",
    "pyyaml>=6.0",
    "requests>=2.28",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "docker>=7.0",
    "ruff>=0.1",
]

[tool.setuptools.packages.find]
where = ["src", "."]
include = ["bdi_llm*", "scripts*"]
exclude = ["tests*"]

[tool.ruff]
line-length = 100

[tool.ruff.lint]
select = ["E", "W", "F", "I", "C90", "UP", "B"]

[tool.ruff.lint.mccabe]
max-complexity = 15

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
